[Sam] Refactor Code Refactor the code into different classes Fixes #65
