+++ src/main/java/seedu/duke/Project.java
@@ -34,12 +34,8 @@ public class Project {
         return false;
     }
 
-
     public boolean checkResourceExistsByIndex(int idx) {
-        if (idx >= resources.size() || idx < 0) {
-            return false;
-        }
-        return true;
+        return !(idx >= resources.size() || idx < 0);
     }
 
     @Override
