+++ src/main/java/seedu/duke/resource/ResourceManager.java
@@ -2,5 +2,4 @@ package seedu.duke.resource;
 
 public class ResourceManager {
 
-
 }
