+++ src/test/java/seedu/duke/DukeTest.java
@@ -4,17 +4,18 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
+import seedu.duke.ui.MainUi;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 
 class DukeTest {
-    String dukeStandardHeading = Duke.LOGO_STRING + "\n"
-            + Duke.PROJECT_TEAM_ID + "\n"
-            + Duke.APP_NAME_AND_VERSION + "\n"
-            + Duke.HOW_TO_GET_HELP + "\n"
-            + Duke.SIGNAL_FOR_USER_TO_INPUT;
+    String dukeStandardHeading = MainUi.LOGO_STRING + "\n"
+            + MainUi.PROJECT_TEAM_ID + "\n"
+            + MainUi.APP_NAME_AND_VERSION + "\n"
+            + MainUi.HOW_TO_GET_HELP + "\n"
+            + MainUi.SIGNAL_FOR_USER_TO_INPUT;
 
     @Test
     public void dummyTest() {
@@ -36,15 +37,15 @@ class DukeTest {
 
         String targetString = dukeStandardHeading
                 + "The resource is added into the new project \"CS2113 Group Project\".\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "Here is the list of all project(s) and it's resource(s)!\n"
                 + "--------------------------------------------------------" + "\n"
                 + "Project 1: CS2113 Group Project\n"
                 + "Resource(s):\n"
                 + "1): https://ay2021s2-cs2113-w10-3.github.io/tp/ (Description: Team Project)\n"
                 + "--------------------------------------------------------\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
-                + Duke.EXIT_MESSAGE + "\n";
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.EXIT_MESSAGE + "\n";
 
         assertEquals(newOutputStream.toString(), targetString);
 
@@ -67,9 +68,9 @@ class DukeTest {
 
         String targetString = dukeStandardHeading
                 + "The resource is added into the new project \"CS2113 Group Project\".\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "The resource is added to the existing project \"CS2113 Group Project\".\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "Here is the list of all project(s) and it's resource(s)!\n"
                 + "--------------------------------------------------------" + "\n"
                 + "Project 1: CS2113 Group Project\n"
@@ -77,8 +78,8 @@ class DukeTest {
                 + "1): https://ay2021s2-cs2113-w10-3.github.io/tp/ (Description: Team Project)\n"
                 + "2): other website\n"
                 + "--------------------------------------------------------\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
-                + Duke.EXIT_MESSAGE + "\n";
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.EXIT_MESSAGE + "\n";
 
         assertEquals(newOutputStream.toString(), targetString);
 
@@ -101,17 +102,17 @@ class DukeTest {
 
         String targetString = dukeStandardHeading
                 + "The resource is added into the new project \"CS2113 Group Project\".\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "The resource of the project \"CS2113 Group Project\" is overwritten.\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "Here is the list of all project(s) and it's resource(s)!\n"
                 + "--------------------------------------------------------" + "\n"
                 + "Project 1: CS2113 Group Project\n"
                 + "Resource(s):\n"
                 + "1): https://ay2021s2-cs2113-w10-3.github.io/tp/ (Description: Team Project for CS2113)\n"
                 + "--------------------------------------------------------\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
-                + Duke.EXIT_MESSAGE + "\n";
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.EXIT_MESSAGE + "\n";
 
         assertEquals(newOutputStream.toString(), targetString);
 
@@ -135,9 +136,9 @@ class DukeTest {
 
         String targetString = dukeStandardHeading
                 + "The resource is added into the new project \"CS2113 Group Project\".\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "The resource is added to the existing project \"CS2113 Group Project\".\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "Here is the list of all project(s) and it's resource(s)!\n"
                 + "--------------------------------------------------------" + "\n"
                 + "Project 1: CS2113 Group Project\n"
@@ -145,10 +146,10 @@ class DukeTest {
                 + "1): https://ay2021s2-cs2113-w10-3.github.io/tp/ (Description: Team Project for CS2113)\n"
                 + "2): https://nus-cs2113-ay2021s2.github.io/website/admin/tp-expectations.html (Description: tp Website)\n"
                 + "--------------------------------------------------------\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "Resource is not found. Please enter a valid index. \n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
-                + Duke.EXIT_MESSAGE + "\n";
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.EXIT_MESSAGE + "\n";
 
         assertEquals(newOutputStream.toString(), targetString);
 
@@ -172,9 +173,9 @@ class DukeTest {
 
         String targetString = dukeStandardHeading
                 + "The resource is added into the new project \"CS2113 Group Project\".\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "The resource is added to the existing project \"CS2113 Group Project\".\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "Here is the list of all project(s) and it's resource(s)!\n"
                 + "--------------------------------------------------------" + "\n"
                 + "Project 1: CS2113 Group Project\n"
@@ -182,10 +183,10 @@ class DukeTest {
                 + "1): https://ay2021s2-cs2113-w10-3.github.io/tp/ (Description: Team Project for CS2113)\n"
                 + "2): https://nus-cs2113-ay2021s2.github.io/website/admin/tp-expectations.html (Description: tp Website)\n"
                 + "--------------------------------------------------------\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "Project is not found ... \n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
-                + Duke.EXIT_MESSAGE + "\n";
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.EXIT_MESSAGE + "\n";
 
         assertEquals(newOutputStream.toString(), targetString);
 
@@ -210,9 +211,9 @@ class DukeTest {
 
         String targetString = dukeStandardHeading
                 + "The resource is added into the new project \"CS2113 Group Project\".\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "The resource is added to the existing project \"CS2113 Group Project\".\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "Here is the list of all project(s) and it's resource(s)!\n"
                 + "--------------------------------------------------------" + "\n"
                 + "Project 1: CS2113 Group Project\n"
@@ -220,17 +221,17 @@ class DukeTest {
                 + "1): https://ay2021s2-cs2113-w10-3.github.io/tp/ (Description: Team Project for CS2113)\n"
                 + "2): https://nus-cs2113-ay2021s2.github.io/website/admin/tp-expectations.html (Description: tp Website)\n"
                 + "--------------------------------------------------------\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "The resource is deleted from the project \"CS2113 Group Project\".\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
                 + "Here is the list of all project(s) and it's resource(s)!\n"
                 + "--------------------------------------------------------" + "\n"
                 + "Project 1: CS2113 Group Project\n"
                 + "Resource(s):\n"
                 + "1): https://nus-cs2113-ay2021s2.github.io/website/admin/tp-expectations.html (Description: tp Website)\n"
                 + "--------------------------------------------------------\n"
-                + Duke.SIGNAL_FOR_USER_TO_INPUT
-                + Duke.EXIT_MESSAGE + "\n";
+                + MainUi.SIGNAL_FOR_USER_TO_INPUT
+                + MainUi.EXIT_MESSAGE + "\n";
 
         assertEquals(newOutputStream.toString(), targetString);
 
@@ -253,7 +254,7 @@ class DukeTest {
                 + "exit: Exits the program.\n"
                 + "------------------------------------------------------------------------\n\n";
 
-        Duke.listAllCommands();
+        MainUi.listAllCommands();
         assertEquals(newOutputStream.toString(), helpExpectedOutput);
 
         System.setOut(System.out);
@@ -271,7 +272,7 @@ class DukeTest {
         Duke.main(null);
 
         String helpExpectedOutput = dukeStandardHeading
-                + Duke.EXIT_MESSAGE + "\n";
+                + MainUi.EXIT_MESSAGE + "\n";
 
         assertEquals(newOutputStream.toString(), helpExpectedOutput);
